fix(ui): make preview indicator offset type explicit

+++ android-app/app/src/main/java/io/github/xgl34222220/luoshu/ui/studio/StudioCompositePreviewCompact.kt
@@ -192,7 +192,7 @@ private fun CompactScenarioSelector(
     val scenarios = CompactPreviewScenario.entries
     val selectedIndex = scenarios.indexOf(selected).coerceAtLeast(0)
     val indicatorOffset by animateDpAsState(
-        targetValue = 78.dp * selectedIndex,
+        targetValue = 78.dp * selectedIndex.toFloat(),
         animationSpec = tween(220, easing = FastOutSlowInEasing),
         label = "previewIndicatorOffset",
     )
